app/requirements.txt
Flask
Flask-SQLAlchemy
flask-migrate
flask_marshmallow
python-dotenv
pymysql
bcrypt
marshmallow-sqlalchemy
Flask-Login